More recommended changes made & controller updated
