+++ .gitignore
@@ -19,6 +19,8 @@ npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
 
+#file upload
+/server/uploads
 
 #TODO add .env to your gitignore
 __snapshots__
